scraping/management/commands/scrape_tournaments.py
from django.core.management.base import BaseCommand
from django.utils import timezone
import requests
from bs4 import BeautifulSoup
import time
import logging

logger = logging.getLogger(__name__)


class Command(BaseCommand):
    """
    法的リスク軽減型スクレイピングコマンド
    
    使用方法:
    python manage.py scrape_tournaments
    """
    
    help = '公的機関から大会情報を自動収集します（法的リスク軽減型）'
    
    def add_arguments(self, parser):
        parser.add_argument(
            '--dry-run',
            action='store_true',
            help='実際には保存せず、収集される情報のみを表示'
        )
        parser.add_argument(
            '--source',
            type=str,
            choices=['jba', 'prefectures'],
            default='prefectures',
            help='スクレイピング対象の選択'
        )
    
    def handle(self, *args, **options):
        """メインの処理"""
        self.stdout.write(
            self.style.SUCCESS('法的リスク軽減型スクレイピングを開始します...')
        )
        
        # 実装方針:
        # 1. robots.txtの確認とアクセス制限の遵守
        # 2. 公的機関（各都道府県バドミントン協会）のみを対象
        # 3. 個人情報を含まない公開情報のみを収集
        # 4. サーバー負荷軽減のため適切な間隔でアクセス
        
        if options['dry_run']:
            self.stdout.write(
                self.style.WARNING('DRY RUN モード: データは保存されません')
            )
        
        source = options['source']
        
        if source == 'prefectures':
            self._scrape_prefecture_associations()
        elif source == 'jba':
            self._scrape_jba_tournaments()
        
        self.stdout.write(
            self.style.SUCCESS('スクレイピング処理が完了しました')
        )
    
    def _scrape_prefecture_associations(self):
        """都道府県バドミントン協会からのデータ収集"""
        self.stdout.write('都道府県協会からの情報収集を開始...')
        
        # TODO: 実際の実装では、以下の点を考慮する必要があります:
        # 1. 各都道府県協会の公式サイトのrobots.txtを確認
        # 2. 利用規約で自動アクセスが禁止されていないか確認
        # 3. 著作権を侵害しない範囲での情報取得
        # 4. 個人情報（担当者名、電話番号等）の除外
        
        # サンプル実装（実際のサイトには接続しません）
        prefecture_urls = [
            # 'https://example-badminton-association.jp/tournaments',
            # 注意: 実際のURLは法的確認後に追加
        ]
        
        for url in prefecture_urls:
            self.stdout.write(f'処理中: {url}')
            # time.sleep(2)  # サーバー負荷軽減のため2秒間隔
            
            # 実際の実装では以下の処理を行います:
            # 1. robots.txtの確認
            # 2. 適切なUser-Agentの設定
            # 3. レスポンシブな間隔でのアクセス
            # 4. エラーハンドリング
            
        self.stdout.write('都道府県協会からの収集完了')
    
    def _scrape_jba_tournaments(self):
        """日本バドミントン協会公式情報の収集"""
        self.stdout.write('日本バドミントン協会からの情報収集を開始...')
        
        # TODO: JBA公式サイトからの情報収集
        # 注意: 事前に利用規約とrobots.txtの確認が必要
        
        self.stdout.write('JBAからの収集完了')
    
    def _check_robots_txt(self, base_url):
        """robots.txtの確認"""
        robots_url = f"{base_url}/robots.txt"
        try:
            response = requests.get(robots_url, timeout=10)
            if response.status_code == 200:
                # robots.txtの解析ロジックをここに実装
                return True
        except Exception as e:
            logger.warning(f"robots.txt確認エラー: {e}")
        return False
    
    def _save_tournament_data(self, tournament_data, dry_run=False):
        """収集したデータの保存"""
        if dry_run:
            self.stdout.write(f"DRY RUN: {tournament_data}")
            return
        
        # 実際のデータ保存ロジック
        # Tournament.objects.create(**tournament_data)
        pass