requirements.txt
Django>=4.2.0,<5.0.0
psycopg2-binary>=2.9.0
python-decouple>=3.8
beautifulsoup4>=4.12.0
requests>=2.31.0
celery>=5.3.0
redis>=4.6.0
django-crispy-forms>=2.0
crispy-bootstrap5>=0.7
pillow>=10.0.0